Added code to compute lattice properties.  Also did some refactoring, mostly of the lattice package.

+++ src/lattices/Lattice.java
@@ -33,7 +33,7 @@ public interface Lattice {
     
     void setDimension(int n);
     
-    double getDimension();
+    int getDimension();
     
     /**
      * Return the generator matrix for this latttice
